Simplify homepage to an essay; merge research and projects into /work
Homepage drops the Selected work and Current direction sections in favor
of four paragraphs under the brand hero. Research and projects combine
into a single /work page: a blog-style research list (date rail, serif
titles, hairlines) followed by product-style project blocks with
at-a-glance spec strips, modeled on thinkingmachines.ai's blog index and
Inkling page. The research note becomes a centered essay with title,
byline, deck, and inline contents; the sticky TOC rail is gone. Old
routes 307-redirect to /work; nav, palette, and llms.txt updated.
