+++ app/page.tsx
@@ -1,14 +1,8 @@
 import Link from "next/link"
 import { SiteFooter } from "@/components/site-footer"
 import { SiteHeader } from "@/components/site-header"
-import { getLucciRepositories } from "@/lib/github"
 
-export const revalidate = 3600
-
-export default async function HomePage() {
-  const repositories = await getLucciRepositories()
-  const selected = repositories.slice(0, 3)
-
+export default function HomePage() {
   return (
     <div className="shell" id="top">
       <SiteHeader />
@@ -16,7 +10,7 @@ export default async function HomePage() {
         <section className="hero" aria-labelledby="hero-title">
           <h1 className="hero-wordmark" id="hero-title">Lucci Labs</h1>
           <p className="hero-tagline">Studying intelligent systems by building them.</p>
-          <Link className="home-announcement" href="/research/context-as-infrastructure">
+          <Link className="home-announcement" href="/work/context-as-infrastructure">
             <span className="pill"><span className="status-dot" />New</span>
             <span className="announcement-text">Context as infrastructure — our latest research note</span>
           </Link>
@@ -25,44 +19,30 @@ export default async function HomePage() {
         <section className="home-intro" aria-label="About Lucci Labs">
           <p>
             Lucci Labs is an independent research studio working on tool-using agents,
-            local inference, and the interfaces around them. We formalize a question,
-            build the smallest system that tests it, and publish what the work reveals.
+            local inference, and the interfaces around them. We are interested in what
+            makes intelligent systems dependable in practice: how they receive context,
+            how they act on it, and how the people using them stay oriented to what the
+            system knows and does.
           </p>
-          <p>The work is open source by default.</p>
-        </section>
-
-        <section aria-labelledby="work-title">
-          <div className="section-head">
-            <h2 id="work-title">Selected work</h2>
-          </div>
-          <div className="work-list">
-            <Link className="work-item" href="/research/context-as-infrastructure">
-              <h3 className="work-title">Context as infrastructure</h3>
-              <p className="work-copy">
-                A practical model for separating task framing, evidence, state, and durable
-                memory in tool-using agent systems.
-              </p>
-              <p className="work-meta">Research note · 2026</p>
-            </Link>
-            {selected.map((repository) => (
-              <a className="work-item" href={repository.url} key={repository.name} target="_blank" rel="noreferrer">
-                <h3 className="work-title">{repository.name}<span className="arrow">↗</span></h3>
-                <p className="work-copy">{repository.description}</p>
-                <p className="work-meta">{repository.language || "Open source"} · GitHub</p>
-              </a>
-            ))}
-          </div>
-        </section>
-
-        <section className="research-note" aria-labelledby="direction-title">
-          <p className="label">Current direction</p>
-          <div>
-            <h2 id="direction-title">We build small systems to interrogate large questions.</h2>
-            <p>
-              The work moves between theory and implementation—formalizing a question, then
-              building the minimal system that shows where the idea holds and where it breaks.
-            </p>
-          </div>
+          <p>
+            The method is consistent. We formalize a question, build the smallest system
+            that can test it, and publish what the work reveals — including where the idea
+            holds and where it breaks. Theory and implementation move together; neither is
+            allowed to drift far from the other.
+          </p>
+          <p>
+            The current direction centers on context as infrastructure: treating the
+            information an agent receives as versioned, observable system components
+            rather than a prompt assembled once at the beginning of a run. The research
+            notes and the instruments built alongside them are collected under{" "}
+            <Link className="quiet-link" href="/work">Work</Link>.
+          </p>
+          <p>
+            Everything is open source by default. The code lives on{" "}
+            <a className="quiet-link" href="https://github.com/lucci-xyz" target="_blank" rel="noreferrer">GitHub</a>,
+            and the lab reads its mail at{" "}
+            <a className="quiet-link" href="mailto:contact@luccilabs.xyz">contact@luccilabs.xyz</a>.
+          </p>
         </section>
       </main>
       <SiteFooter />
