+++ components/command-palette.tsx
@@ -60,8 +60,7 @@ export function CommandPalette() {
   const commands = useMemo<Command[]>(() => {
     const navigate: Command[] = [
       { id: "nav-home", group: "Navigate", title: "Home", meta: "Page", keywords: "index start lucci labs", run: () => router.push("/") },
-      { id: "nav-research", group: "Navigate", title: "Research", meta: "Page", keywords: "notes papers publications writing", run: () => router.push("/research") },
-      { id: "nav-projects", group: "Navigate", title: "Projects", meta: "Page", keywords: "open source github repositories instruments", run: () => router.push("/projects") },
+      { id: "nav-work", group: "Navigate", title: "Work", meta: "Page", keywords: "research notes papers publications projects open source github instruments", run: () => router.push("/work") },
       { id: "nav-docs", group: "Navigate", title: "Docs", meta: "Page", keywords: "documentation quickstart sdk api", run: () => router.push("/docs") },
     ]
     const research: Command[] = researchItems.map((item) => ({
@@ -72,7 +71,7 @@ export function CommandPalette() {
       description: item.description,
       keywords: `${item.format} ${item.category} ${item.description}`,
       forthcoming: !item.slug,
-      run: item.slug ? () => router.push(`/research/${item.slug}`) : undefined,
+      run: item.slug ? () => router.push(`/work/${item.slug}`) : undefined,
     }))
     const actions: Command[] = [
       {
