+++ components/scroll-reveal.tsx
@@ -8,14 +8,9 @@ const SELECTOR = [
   ".hero",
   ".home-intro",
   ".section-head",
-  ".work-item",
-  ".research-note",
-  ".about",
   ".page-hero",
-  ".filter-bar",
   ".index-entry",
   ".project-entry",
-  ".page-note",
   ".docs-content section",
   ".doc-head",
   ".document-body > section",
